# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Leita Judge System API Documentation",
        "title": "Leita API Docs",
        "contact": {},
        "version": "2.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/problem/run/{problemId}": {
            "post": {
                "description": "Execute code against provided test cases for testing purposes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Problem"
                ],
                "summary": "Run code with test cases",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Problem ID",
                        "name": "problemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Code and test cases",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.RunProblemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.RunProblemResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.RunProblemResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.RunProblemResponse"
                            }
                        }
                    }
                }
            }
        },
        "/problem/submit/{problemId}": {
            "post": {
                "description": "Submit code for a specific problem to be judged.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Problem"
                ],
                "summary": "Submit a problem solution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Problem ID",
                        "name": "problemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Solution code and metadata",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.SubmitProblemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.SubmitProblemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.SubmitProblemResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.SubmitProblemResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Limit": {
            "type": "object",
            "properties": {
                "memory": {
                    "type": "integer"
                },
                "time": {
                    "type": "integer"
                }
            }
        },
        "entity.RunProblemRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "limit": {
                    "$ref": "#/definitions/entity.Limit"
                },
                "testCases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.TestCase"
                    }
                }
            }
        },
        "entity.RunProblemResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                }
            }
        },
        "entity.SubmitProblemRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "limit": {
                    "$ref": "#/definitions/entity.Limit"
                },
                "submitId": {
                    "type": "integer"
                }
            }
        },
        "entity.SubmitProblemResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "usedMemory": {
                    "type": "integer"
                },
                "usedTime": {
                    "type": "integer"
                }
            }
        },
        "entity.TestCase": {
            "type": "object",
            "properties": {
                "input": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                }
            }
        }
    }
}